refactor(dashboard): extract 7 domain modules from api/dashboard.ts (A1 phase 1)
Split dashboard.ts (4243 -> 3713 lines) into domain files. dashboard.ts
retained as facade (re-export) so consumers importing from './api/dashboard'
are unchanged.

New modules:
- dashboard-shared.ts: cross-domain types (TelemetryFreshnessMetadata,
  DashboardSurfaceEnvelope, TelemetryCoverageGap, DashboardFeedMetadata)
- dashboard-fusion.ts, dashboard-agent.ts
- dashboard-mission.ts, dashboard-keeper-trajectory.ts
- dashboard-keeper-tool-stats.ts, dashboard-keeper-tool-calls.ts

typecheck green (tsc --noEmit). No consumer edits.

11 domains remain (logs/execution/governance/runtime/keeper-cost/goals/
tools-prompts/keeper-config/turn-records/telemetry/misc) for phase 2.

Ref: planning/masc-bplus-verified-checklist-2026-06-24.md (A1)
